Upgrade to Titanium SDK 12.1.2.gradle
Also remove hyperloop and Crashlytics hacks because this experiment
didn't work out. Whilst crashlytics sort of works hyperloop hacks
significantly complicate the build process and is it worh it?

The final nail in the coffin is that most of our hard to diagnose issues
happen when the phone is out of range so we don't see the logs anyway.

Will need to revisit a way to get offline logs.


Former-commit-id: 9970dd97a898fafeb716b0951c8ab2f427085382