Fix duplicate created whilst editing taxon
The issue was that in the following sequence of events:
- create new sample
- sample starts uploading
- edit sample before uploaded
- save edited sample before uploading has finished

Then the `serverSampleId` field is set on an in memory copy
of the sample record. This is then saved, which recreates the
record that was mean't to be deleted.

This commit fixes this by first trying to reload the sample, and
if it no longer exists, looking up the duplicated record.

`originalSampleId` is no longer set to `null` as a result but
I can't think of a scenario where this an issue.

Once the original sample is deleted this id will no longer point
to a valid sample.

There are two places where it is referenced:
   1. When uploading a sample and the original record is already
      destroyed.
   2. When saving an edit, in which case the original record is looked
      up and destroyed.


Former-commit-id: 65783569e8f8d9c1bfb4316615d84096ddc9353f